common: Add a compile-time option to use OpenSSL's async interface for hashing
This option is not a good idea to turn on in general. If using the standard software-based
openssl hashing implementations this will significantly reduce the performance of programs such as
radosgw that make heavy use of hashing. This is because all communication with openssl is serialised
through a single thread, which is a significant bottleneck.

However in the case where a hardware offload engine for OpenSSL which supports hashing operations in
asynchronous mode is installed then this code could dramatically reduce cpu usage during these
operations by allowing the hardware offload engine to pause jobs whilst waiting for responses from the
hardware engine. Paused jobs will still be polled by the worker thread that communicates with openssl,
but all other threads waiting on the result will wait on condition_variables.

Signed-off-by: James Weaver <james.barrett@bbc.co.uk>